Makes things work and adds new features
The AudioConnectClient and AudioConnectClientHandler classes were refactored, fixed,
and combined into AudioConnectClient which handles all network communication to and
from the AudioConnect web service.
PlayerScheduler class was added to manage schedulers with a balanced load of players
to be tracked rather than using PlayerMoveEvent which fires too often and is restricted to
player movement.
ConnectCommand class was added to allow players to receive in chat their unique URL
to connect to the AudioConnect web service.
The ListCommand was refactored to improve the list display and understandability in chat.
The ReloadCommand was updated to also restart the client and reattempt a connection
with the AudioConnect web service.